Rearrange a lot of code
git.py was getting extremely crowded and hard to navigate. Split
everything up so that we have some sort of logical organization.
